build create_name function to nummerate output files
