+++ src/adaptfx/aft_utils.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 import time
-from adaptfx import aft_message, aft_error, aft_warning, aft_message
+import os
+import adaptfx as afx
 nme = __name__
 
 def stat_rounding(number, decimal):
@@ -30,7 +31,74 @@ def timing(start=None):
     else:
         stop = time.perf_counter()
         time_elapsed = stat_rounding((stop - start), 4)
-        aft_message(f'process duration: {time_elapsed} s:', nme, 1)
+        afx.aft_message(f'process duration: {time_elapsed} s:', nme, 1)
+
+def get_abs_path(filename, name):
+    """
+    from filename create absolute path and basename
+    without ".appendix"
+
+    Parameters
+    ----------
+    filename : string
+        filename of instructions
+    name : string
+        logger name
+
+    Returns
+    -------
+    norm_path : string
+        normalised absolute path
+    basename : string
+        absolute path to file without suffix
+
+        
+    """
+    if os.path.isfile(filename): # check if filepath exists
+        if not os.path.isabs(filename): # check if absolute
+            filename = os.path.abspath(filename)
+        # collapse redundant separators get base
+        norm_path = os.path.normpath(filename)
+        abs_path, name = os.path.split(norm_path)
+        base, _ = os.path.splitext(name)
+        basename = os.path.join(abs_path, base)
+    else:
+        afx.aft_error(f'did not find "{filename}"', name)
+    return norm_path, basename
+
+def create_name(basename, suffix):
+    """
+    from basename create name for a file
+    that does not yet exist. Increments numbers
+
+    Parameters
+    ----------
+    basename: string
+        absolute basename of file without suffix
+    suffix : string
+        suffix of the output
+
+    Returns
+    -------
+    new_name : string
+        new name nummerated
+        
+    """
+    # create logfile name and
+    # search for existing filename ...
+    i = 1
+    while os.path.exists(f'{basename}_{i}.{suffix}'):
+        # exponential search if many files exist
+        i *= 2
+    a, b = (i // 2, i)
+    while a+1 < b:
+        c = (a + b) // 2
+        if os.path.exists(f'{basename}_{c}.{suffix}'):
+            a, b = (c, b)
+        else:
+            a, b = (a, c)
+    # ... end of search
+    return f'{basename}_{b}.{suffix}'
 
 def key_reader(all_keys, full_dict, user_keys, algorithm):
     """
@@ -54,7 +122,7 @@ def key_reader(all_keys, full_dict, user_keys, algorithm):
         all keys copied from parameters
         
     """
-    aft_message('loading keys...', nme, 1)
+    afx.aft_message('loading keys...', nme, 1)
     whole_dict = full_dict.copy()
     key_dict = all_keys[algorithm]
 
@@ -63,17 +131,17 @@ def key_reader(all_keys, full_dict, user_keys, algorithm):
             whole_dict[key] = user_keys[key]
         elif key not in user_keys:
             if full_dict[key] == None:
-                aft_error(f'missing mandatory key: "{key}"', nme)
+                afx.aft_error(f'missing mandatory key: "{key}"', nme)
             else:
                 whole_dict[key] = full_dict[key]
 
     for key in user_keys:
         if key not in key_dict and key in full_dict:
-            aft_warning(
+            afx.aft_warning(
                 f'key: "{key}" is not allowed for "{algorithm}"', nme, 0
                 )
         elif key not in full_dict:
-            aft_warning(f'unexpected key: "{key}"', nme, 0)
+            afx.aft_warning(f'unexpected key: "{key}"', nme, 0)
 
     return dict((k, whole_dict[k]) for k in key_dict)
 
@@ -95,7 +163,7 @@ def setting_reader(all_settings, user_settings):
         all settings
         
     """
-    aft_message('loading settings...', nme, 1)
+    afx.aft_message('loading settings...', nme, 1)
     whole_settings = all_settings.copy()
 
     for skey in all_settings:
@@ -104,7 +172,7 @@ def setting_reader(all_settings, user_settings):
     
     for skey in user_settings:
         if skey not in all_settings:
-            aft_warning(f'unexpected setting: "{skey}"', nme, 0)
+            afx.aft_warning(f'unexpected setting: "{skey}"', nme, 0)
 
     return whole_settings
